api/requirements.txt
fastapi
uvicorn
openai
pdfplumber
python-docx
python-dotenv
pydantic
python-multipart
beautifulsoup4
requests
